Added the role management and the institution based data

+++ app/Http/Controllers/PlansController.php
@@ -2,28 +2,36 @@
 
 namespace App\Http\Controllers;
 
-use RealRashid\SweetAlert\Facades\Alert;
-use Auth;
+use App\Constants;
 use DateTime;
-use Exception;
 use Log;
-use Redirect;
 use App\Plan;
 
 use App\Exports\PlanExcelExport;
 use App\Exports\PlanPdfExport;
+use App\User;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
 use Maatwebsite\Excel\Facades\Excel;
 use Jenssegers\Agent\Agent;
 
 class PlansController extends Controller
 {
+    function __construct() {
+        parent::__construct();
+
+        $this->middleware('permission:plan-list|plan-create|plan-delete', ['only' => ['index','show','exportExcel', 'exportPdf']]);
+        $this->middleware('permission:plan-delete', ['only' => ['destroy']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index() {
-        $plans = Plan::where('user_id', '=', Auth::user()->id)->orderBy('created_at', 'desc')->get();
+        $plans = $this->evaluatePlans();
         return view('/plans/planslist', compact('plans'));
     }
 
@@ -36,21 +44,17 @@ class PlansController extends Controller
     public function show(Plan $plan) {
         $agent = new Agent();
 
-        return view('/plans/viewplans', [
-            'plan' => $plan,
-            'inputs' => json_decode($plan->inputs),
-            'result' => json_decode($plan->results),
-            'agent' => $agent
-        ]);
+        if ($this->evaluatePlanSelected($plan)) {
+            return view('/plans/viewplans', [
+                'plan' => $plan,
+                'inputs' => json_decode($plan->inputs),
+                'result' => json_decode($plan->results),
+                'agent' => $agent
+            ]);
+        } else {
+            return Redirect::to('/404');
+        }
     }
-
-    // public function edit(Plan $plan) {
-    //     return view('/plans/viewplans', [
-    //         'plan' => $plan,
-    //         'inputs' => json_decode($plan->inputs),
-    //         'result' => json_decode($plan->results)
-    //     ]);
-    // }
 
     /**
      * Download the specified resource in storage.
@@ -60,18 +64,28 @@ class PlansController extends Controller
      */
     public function exportExcel($id) {
         $plan = Plan::find($id);
-        $date = new DateTime();
-        return Excel::download(new PlanExcelExport($plan), trans('messages.profitability') . '-' . $plan->business_code . '-' . $date->format('d-m-Y-H-i-s') . '.xlsx');
+
+        if ($this->evaluatePlanSelected($plan)) {
+            $date = new DateTime();
+            return Excel::download(new PlanExcelExport($plan), trans('messages.profitability') . '-' . $plan->business_code . '-' . $date->format('d-m-Y-H-i-s') . '.xlsx');
+        } else {
+            return Redirect::to('/404');
+        }
     }
 
     public function exportPdf($id) {
         $plan = Plan::find($id);
-        $date = new DateTime();
 
-        $pdf = new PlanPdfExport($plan);
-        $doc = $pdf->export();
+        if ($this->evaluatePlanSelected($plan)) {
+            $date = new DateTime();
 
-        return $doc->download(trans('messages.profitability') . '-' . $plan->business_code . '-' . $date->format('d-m-Y-H-i-s') . '.pdf');
+            $pdf = new PlanPdfExport($plan);
+            $doc = $pdf->export();
+
+            return $doc->download(trans('messages.profitability') . '-' . $plan->business_code . '-' . $date->format('d-m-Y-H-i-s') . '.pdf');
+        } else {
+            return Redirect::to('/404');
+        }
     }
 
     /**
@@ -81,7 +95,54 @@ class PlansController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Plan $plan) {
-        $plan->delete();
-        return Redirect::to('/plans')->withSuccessMessage(trans('messages.project_deleted'));
+        if ($this->evaluatePlanSelected($plan)) {
+            $plan->delete();
+            return Redirect::to('/plans')->withSuccessMessage(trans('messages.project_deleted'));
+        } else {
+            return Redirect::to('/404');
+        }
+    }
+
+    /**
+     * Understand what plans to load based on the role of the user.
+     */
+    private function evaluatePlans() {
+        if (Auth::user()->roles[0]->id == Constants::ROLE_ADMIN_ID) {
+            $plans = Plan::with('user')->with('user')->orderBy('created_at', 'desc')->get();
+        } else if (Auth::user()->roles[0]->id == Constants::ROLE_INSTITUTION_ID) {
+            $plans = Plan::whereHas('user', function (Builder $query) {
+                    $query->where('user_related_id', '=', Auth::user()->id);
+                })->orWhere('user_id', '=', Auth::user()->id)
+                ->with('user')
+                ->orderBy('created_at', 'desc')
+                ->get();
+        } else {
+            $plans = Plan::where('user_id', '=', Auth::user()->id)->orderBy('created_at', 'desc')->get();
+        }
+
+        return $plans;
+    }
+
+    /**
+     * Check if the plan selected fits to the roles of the users.
+     */
+    private function evaluatePlanSelected($plan) {
+        if (Auth::user()->roles[0]->id == Constants::ROLE_ADMIN_ID) {
+            return true;
+        } else if (Auth::user()->roles[0]->id == Constants::ROLE_INSTITUTION_ID) {
+            $userIds = array();
+            array_push($userIds, Auth::user()->id);
+            $users = User::where('user_related_id', '=', Auth::user()->id)->get();
+
+            foreach ($users as $user) {
+                array_push($userIds, $user->id);
+            }
+
+            if (in_array($plan->user_id, $userIds)) return true;
+            else return false;
+        } else {
+            if ($plan->user_id == Auth::user()->id) return true;
+            else return false;
+        }
     }
 }
